Added support for generation of API Overview section
For Swagger APIs that have a verbose description that is formatted using
markdown, this code splits the markdown out, converts it to html, and adds
a new "API Overview" tab for the content.  If there is an API overview tab,
which there only is for local APIs at this moment, this content shows up,
otherwise it is empty for now.

This revision also overrides some of the clarity defaults for colors so that
it matches the vSphere API Explorer as requested.

Signed-off-by: Aaron Spear <aspear@vmware.com>