Added additional variables to control display of the filter pane
Updated config.js with new hide* variables that control whether or not the entire
filter pane or particular sections (e.g products) are displayed for the user to
control them.  This is decoupled from the defaultFilters, so the net result is that
it is easily possible to configure the API Explorer such that you have an API explorer
scoped to a particular product with no possibility of changing it to see other products.
Also updated documentation accordingly.  Default behavior is unchanged.

Signed-off-by: Aaron Spear <aspear@vmware.com>