added working support for vRA SSO/token authorization
Right now this was done by simply cloning the previously hard coded vsphere support and
renaming it.  Now, but the vsphere and vra SSO implementations exist in the code, named
appropriately, and you can select which one you want by specifying a new ssoId string
value in the config.js file.

I don't actually think this is the ideal implementation by any means, but it seemed
like the most simple path to be able to enable vRA SSO and vSphere SSO to live together
 in the same code base short term.  I think a better implementation would allow
individual APIs to specify the ssoId that they want to use and then the particular
SSO that appears when you hit the log in button would depend on which API you were using.
It should be possible for them to be used simultaneously, which would also require using
 of a different cookie for each one.

Signed-off-by: Aaron Spear <aspear@vmware.com>