Added optional support for basic auth for all APIs
Optionally a user can now specific in config.js that they want to use basic
auth for all APIs. When they do this, the username and password they specify
are passed to the swagger widget for any API they open, with the authorization
having the name specified in window.config.swaggerAuthName.  This allows
providing a default value for any swagger basic auth of that name.

This is version 0.36.  Also updated apixlocal tools to version 0.6

Signed-off-by: Aaron Spear <aspear@vmware.com>